% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{BM}
\alias{BM}
\title{Massive Brunner-Munzel tests}
\usage{
BM(lesmat, behavior)
}
\arguments{
\item{lesmat}{matrix of voxels}

\item{behavior}{vector of behavior}
}
\value{
list of statistic and degrees of freedom
}
\description{
Takes a binary matrix of voxels and a vector of behavior 
and runs Brunner-Munzel tests on each voxel.
This function is not compiled and is slow.
}
\author{
Dorian Pustina
}
