% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BMfast}
\alias{BMfast}
\title{Fast Brunner-Munzel tests (v1)}
\usage{
BMfast(X, y)
}
\arguments{
\item{X}{binary matrix ov voxlels (columns) for all
subjects (rows)}

\item{y}{vector of behavioral scores.}
}
\value{
List with two vectors:
- statistic - BM values
- dfbm - degrees of freedom
}
\description{
Takes a binary matrix of voxels and a vector of behavior
and runs Brunner-Munzel tests on each voxel.
This is a fast function, but may produce infinite values
for perfectly separated group. Use BMfast2 which avoids
this problem.
}
\author{
Dorian Pustina

/@export
}
