% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BMfast2}
\alias{BMfast2}
\title{Fast Brunner-Munzel tests (v2)}
\usage{
BMfast2(X, y, computeDOF = TRUE)
}
\arguments{
\item{X}{binary matrix ov voxlels (columns) for all
subjects (rows)}

\item{y}{vector of behavioral scores.}

\item{computeDOF}{(true) chooses whether to compute degrees
of freedom. Set to false to save time during permutations.}
}
\value{
List with two vectors:
- statistic - BM values
- dfbm - degrees of freedom
}
\description{
Takes a binary matrix of voxels and a vector of behavior
and runs Brunner-Munzel tests on each voxel.
This is a fast function that corrects for infinite values
with a similar approach as the nparcomp package.
}
\author{
Dorian Pustina
}
