% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BMperm}
\alias{BMperm}
\title{Fast Brunner-Munzel tests (v2) with permutations}
\usage{
BMperm(X, y, computeDOF = TRUE, npermBM = 20000L, alternative = 1L)
}
\arguments{
\item{X}{binary matrix ov voxlels (columns) for all
subjects (rows)}

\item{y}{vector of behavioral scores.}

\item{computeDOF}{(default true) chooses whether to compute
degrees of freedom. Set to false to save time during permutations.}

\item{alternative}{(default 1) integer to select the tail of
pvalues. 1-greater, 2-less, 3-two.sided}

\item{nperm}{(default 20000) number of permutations to run at each
voxel}
}
\value{
List with these objects:
- statistic - BM values
- dfbm - degrees of freedom
- pvalue - permutation-based probability value
}
\description{
Takes a binary matrix of voxels and a vector of behavior
and runs Brunner-Munzel tests on each voxel.
This is a fast function that corrects for infinite values
with a similar approach as the nparcomp package.
It calculates p-values by running permutations of each
voxel and using the ratio of times the real BM score
exceeds the permuted BM score.
}
\author{
Dorian Pustina
}
