% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_ttest.R
\name{checkAssumptions_ttest}
\alias{checkAssumptions_ttest}
\title{checkAssumptions_ttest}
\usage{
checkAssumptions_ttest(lesmat, behavior, assumptionThreshold = 0.05,
  showInfo = TRUE, ...)
}
\arguments{
\item{lesmat}{binary matrix (0/1) of voxels (columns)
and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{assumptionThreshold}{(default=0.05) threshold to
establish whether tests pass or fail. This threshold is
applied at each voxel individually, no corretion for
multiple comparison is applied.}

\item{showInfo}{logical (default=TRUE), display info messages}

\item{...}{arguments that are passed by upstream functions}
}
\value{
List of objects returned:
\itemize{
\item\code{failVarianceTest} - vector of logical values
   marking voxels that have different variance of behavioral
   scores in lesioned and non-lesioned individuals. Obtained
   using the \code{var.test} function.
\item\code{failNormalityTest} - vector of logical values
   marking voxels with abnormal distribution of behavioral
   scores either in lesioned or non-lesioned individuals.
   Obtained with the \code{shapiro.test} function.
}
}
\description{
Routine to test statistical assumptions are met
at each voxel for t-tests
}
\author{
Dorian Pustina
}
