% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFilenameHeaders.R
\name{checkFilenameHeaders}
\alias{checkFilenameHeaders}
\title{Check headers of files stored on disk}
\usage{
checkFilenameHeaders(files, showError = TRUE)
}
\arguments{
\item{files}{character vector of filenames}

\item{showError}{logical whether to show an error (True) or
to return a boolean instead. Returned
values are True=pass,False=Fail}
}
\value{
logical if the test was successful or not
}
\description{
Function to check that all filenames in a vector
point to existing files with the same resolution,
orientation, size, and origin.
}
\author{
Dorian Pustina
}
