% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMask.R
\name{checkMask}
\alias{checkMask}
\title{checkMask}
\usage{
checkMask(lesions.list, mask)
}
\arguments{
\item{mask}{antsImage of mask to check}

\item{lesion.list}{list of antsImages or character vector
of filenames}
}
\value{
Nothing returned, funciton stops with error if
mask is not in same space as images in lesions.list
}
\description{
Function to check if mask is in the same space as inputs
}
\author{
Dorian Pustina
}

