% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolds.R
\name{.createFolds}
\alias{.createFolds}
\title{Create folds for k-fold validation}
\usage{
.createFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{split sample by balancing y}

\item{k}{number of folds}

\item{list}{logical whether to return folds in a list}

\item{returnTrain}{logical whether to return training indices (T)
or the test samples (F)}
}
\description{
Used to create balanced folds with respect to y input
for k-fold validation.
}
\author{
Caret Package
}
