% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLesionLoad.R
\name{getLesionLoad}
\alias{getLesionLoad}
\title{getLesionLoad}
\usage{
getLesionLoad(lesions.list, parcellation, label = NA, mask = NA,
  binaryCheck = F, keepAllLabels = F, minSubjectPerLabel = "10\%")
}
\arguments{
\item{lesions.list}{list of antsImages or filenames.
Must be binary (0 and 1 values)}

\item{parcellation}{parcellation volume as antsImage or filename.}

\item{label}{You can select the specific label number to compute
the lesion load from. Default is all labels.}

\item{mask}{If a mask is specified (antsImage or filename)
lesioned voxels outside the mask are ignored.}

\item{binaryCheck}{check lesion files are binary (0 and 1). Will
output an error if lesion files are not binary.}

\item{keepAllLabels}{- by default we remove unaffected and 0
labels. Set this to True to keep all labels.}

\item{minSubjectPerLabel}{- minimum number of subjects a parcel
must be lesioned to be kept among returned parcels}
}
\value{
Matrix of lesion loads (0 to 1). Each column is a single
parcel and each row a single subject. Parcel numbers are placed
as column names. 1 means 100% lesioned.
}
\description{
Computes lesion loads from a series of images. A parcellation
image (or simple mask) is required to define the regions from
thish to compute the lesion load.
}
\author{
Dorian Pustina
}
