% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLesionSize.R
\name{getLesionSize}
\alias{getLesionSize}
\title{getLesionSize}
\usage{
getLesionSize(lesions.list)
}
\arguments{
\item{lesions.list}{List of antsImages. For proper
measurement, images must be binary.}
}
\value{
vector of lesion sizes
}
\description{
Compute lesion sizes from a list of antsImages.
}
\author{
Dorian Pustina
}

