% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueLesionPatches.R
\name{getUniqueLesionPatches}
\alias{getUniqueLesionPatches}
\title{Unique Lesion Patches}
\usage{
getUniqueLesionPatches(lesions.list, mask = NA, returnPatchMatrix = F,
  thresholdPercent = 0.1, binaryCheck = F, showInfo = T)
}
\arguments{
\item{lesions.list}{list of antsImages (faster) or filenames (slower)}

\item{mask}{a mask image to restrict the search for patches. Will be
automatically calculated if not provided (voxels > 0 in at
least one subject)}

\item{returnPatchMatrix}{logical, should the matrix of patches
be returned}

\item{thresholdPercent}{voxels with lesions in less than X percent of
subjects will not be considered (default 10\%)}

\item{binaryCheck}{logical, if True images will be verified that are binary}

\item{showInfo}{logical indicating whether to display information (default=T)}
}
\value{
List of objects named as follows:
   patchimg - antsImage with the patch number each voxels belongs to
   patchimg.samples - antsImage mask with a single voxel per patch
   patchimg.size - antsImage with the patch size at each voxel
   patchimg.mask - the mask within which the function looked for patches
   npatches - number of unique patches in the image
   nvoxels - total number of lesioned voxels, computed as all voxels
      lesioned in at least one subject, within the specified mask
   patchvoxels - vector of voxel count for each patch
   patchvolumes - vector of volume size for each patch
   patchmatrix  - matrix of patches


 @examples
 \dontrun{
 files = Sys.glob('/data/jag/VLSM/*.nii.gz')
 lesions = imageFileNames2ImageList(files)

 patches = getUniqueLesionPatches(files) # slower
 patches = getUniqueLesionPatches(lesions) # faster
 avgles = antsAverageImages(lesions)
 mask = thresholdImage(avgles, 0.1, Inf)
 patches = getUniqueLesionPatches(lesions, mask=mask) # even faster
 }
}
\description{
Compute uniqe patches of voxels with the same
pattern of lesions in all subjects. Useful to understand
the amount of unique information in a lesion dataset.
I.e., your dataset can have 1000 subjects but still have
little spatial discrimination because all lesions are
similar
}
\author{
Dorian Pustina
}

