% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesyload_mricron.R
\name{lesyload_mricron}
\alias{lesyload_mricron}
\title{Get inputs from MRIcron/npm file}
\usage{
lesyload_mricron(valfile, imageFolder = NA, returnFilenames = F,
  checkHeaders = T, showInfo = T)
}
\arguments{
\item{valfile}{mricron filename with extention *.val. The function
will search for images in the same folder where valfile
is located, unless you specify \code{imageFolder}.
If any of the files listed in the .val file are not
found in the folder, an error will be displayed.}

\item{imageFolder}{(default=NA) folder to look for the image files}

\item{returnFilenames}{(default=FALSE) By default the function will
load the images in memory to speed up things in lesymap.
This may require too much RAM memory in some cases, and
you may want to use filenames instead, which requires less
memory but is slower in lesymap.}

\item{checkHeaders}{(default=TRUE) Headers will be checked to make sure
all images have the same dimension/origin/resolution, etc.}

\item{showInfo}{(default=TRUE) show information upon successful load}
}
\value{
List with the following information
     lesions - list of antsImages or vector of filenames
     behavior - vector of behavioral scores
}
\description{
Function to load data from a previous analysis
in MRIcron/npm in a ready format for use in lesymap
}
\author{
Dorian Pustina
}
