% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_BMfast.R
\name{lsm_BMfast}
\alias{lsm_BMfast}
\title{lsm_BMfast}
\usage{
lsm_BMfast(lesmat, behavior, permuteNthreshold = 9, alternative = "greater",
  statOnly = F, nperm = 1000, FWERperm = F, v = 1, pThreshold = 0.05,
  showInfo = F, ...)
}
\arguments{
\item{lesmat}{binary matrix (0/1) of voxels (columns)
and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{permuteNthreshold}{(default=9) Voxels lesioned in less than
this number will undergo permutation based thresholding.
See Medina et al 2010.}

\item{alternative}{(default="greater") It is assumed that
healthy voxels (0) have greater behavioral scores. If your
data follow an inverted relationship choose "less" or
"two.sided".}

\item{statOnly}{logical (default=FALSE), skips some computations,
don't use unless you know it's effects}

\item{nperm}{(default=1000) Number of permutations to perform when needed.}

\item{FWERperm}{logical (default=FALSE) whether to perform permutation
based FWER thresholding.}

\item{v}{(default=1) which voxel to record at each permutation, first
or other voxels (i.e., v=10 for 10 highest voxel)}

\item{pThreshold}{(default=0.05) what threshold to use for FWER}

\item{showInfo}{display info messagges when running the function.}

\item{...}{other arguments received from \code{\link{lesymap}}.}
}
\value{
List of objects returned:
\itemize{
\item\code{statistic} - vector of statistical values
\item\code{pvalue} - vector of pvalues
\item\code{zscore} - vector of zscores
\item\code{perm.vector} - (optional) vector of permuted statistics
\item\code{perm.FWERthresh} - (optional) permutation threshold established
from the distribution of \code{perm.vector}
}

List with vectors of statistic, pvalue, zscore.
If FWER is selected, perm.vector and perm.FWERthresh
will also be returned.
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
Brunner-Munzel tests are performed using each column
of the matrix to split the behavioral scores in two
groups. This function relies on a compiled version for
fast processing.
}
\author{
Dorian Pustina
}

