% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_chisq.R
\name{lsm_chisq}
\alias{lsm_chisq}
\title{Chi square tests for symptom mapping}
\usage{
lsm_chisq(lesmat, behavior, YatesCorrect = TRUE, runPermutations = F,
  nperm = 2000, showInfo = TRUE, ...)
}
\arguments{
\item{lesmat}{binary matrix (0/1) of voxels (columns)
and subjects (rows).}

\item{behavior}{vector of behavioral scores (must be binary.}

\item{YatesCorrect}{(default=T) logical whether to use Yates correction.}

\item{runPermutations}{logical (default=FALSE) whether to
use permutation based p-value estimation.}

\item{nperm}{(default=2000) The number of permutations to run.}

\item{showInfo}{display info messagges when running the function.}

\item{...}{other arguments received from \code{\link{lesymap}}.}
}
\value{
List of objects returned:
\itemize{
\item\code{statistic} - vector of statistical values
\item\code{pvalue} - vector of pvalues
\item\code{zscore} - vector of zscores
}
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
The behavior must be a binary vector.
Chi square tests are performed at each voxel. By default
the Yates correction is performed, use \code{correct=FALSE}
if you need to disable it. The behavior must
be a binary vector. Exact p-values can be obtained with permutation
based estimatins.
}
\examples{
{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(1234)
behavior = rbinom(100,1,0.5)
result = lsm_chisq(lesmat, behavior)
}

}
\author{
Dorian Pustina
}
