% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_regres.R
\name{lsm_regres}
\alias{lsm_regres}
\title{lsm_regres}
\usage{
lsm_regres(lesmat, behavior)
}
\arguments{
\item{lesmat}{matrix of voxels (columns) and subjects (raws).}

\item{behavior}{vector of behavioral scores.}
}
\value{
List with vectors of statistic, pvalue, and zscore.
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
Regressions are performed between behavior and each column in the
lesmat matrix.
}
\author{
Dorian Pustina
}

