% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_regresfast.R
\name{lsm_regresfast}
\alias{lsm_regresfast}
\title{Regression tests for symptom mapping (fast)}
\usage{
lsm_regresfast(lesmat, behavior, covariates = NA, FWERperm = F,
  nperm = 1000, v = 1, pThreshold = 0.05, clusterPerm = F,
  mask = NA, voxindx = NA, samplemask = NA,
  clusterPermThreshold = 0.05, showInfo = T, ...)
}
\arguments{
\item{lesmat}{matrix of voxels (columns) and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{covariates}{(default=NA) vector of matrix of covariates.}

\item{FWERperm}{logical (default=FALSE) whether to run permutation
based FWER thresholding.}

\item{nperm}{Number of permutations to perform when needed.}

\item{v}{(default=1) what voxel to record for FWER thresholding.}

\item{pThreshold}{(default=0.05) Voxel-wise threshold.}

\item{clusterPerm}{logical (default=FALSE), whether to perform
permutation based cluster thresholding.}

\item{mask}{(default=NA) antsImage reference mask used for
cluster computations.}

\item{voxindx}{(default=NA) indices of voxels to put in mask}

\item{samplemask}{(default=NA) antsImage used to extract voxels
back in a matrix.}

\item{clusterPermThreshold}{(default=0.05) threshold for cluster
selection after obtaining cluster size distrubution.}

\item{showInfo}{display info messagges when running the function.}

\item{...}{other arguments received from \code{\link{lesymap}}.}
}
\value{
List of objects returned:
\itemize{
\item\code{statistic} - vector of statistical values
\item\code{pvalue} - vector of pvalues
\item\code{zscore} - vector of zscores
\item\code{perm.vector} - (optional) vector of permuted statistics
\item\code{perm.FWERthresh} - (optional) permutation threshold established
from the distribution of \code{perm.vector}
\item\code{perm.clusterThreshold} - (optional) permutation threshold established
from the distribution of \code{perm.vector}

}
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
Regressions are performed between behavior and each column
of the lesmat matrix. Fast function based on compiled code.
}
\examples{
{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(123)
behavior = rnorm(100)
result = lsm_regresfast(lesmat, behavior)
}

}
\author{
Dorian Pustina
}
