% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_sccan.R
\name{lsm_sccan}
\alias{lsm_sccan}
\title{Sparse canonical correlations for symptom mapping.}
\usage{
lsm_sccan(lesmat, behavior, mask, rawStat = F, showInfo = T,
  optimizeSparseness = T, tstamp = "\%H:\%M:\%S", pThreshold = 0.05,
  mycoption = 1, robust = 1, sparseness = 0.045, nvecs = 1,
  cthresh = 150, its = 20, npermsSCCAN = 0, smooth = 0.4, ...)
}
\arguments{
\item{lesmat}{matrix of voxels (columns) and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{mask}{antsImage binary mask to put back voxels in image.}

\item{rawStat}{logical (default=FALSE) whether to skip
the normalization of values and thresholding at 0.1.
If TRUE, the raw voxel weights will be returned.}

\item{showInfo}{logical (default-TRUE) display messages}

\item{optimizeSparseness}{logical (default=TRUE) whether to
run the sparseness optimization routine. If false, the defau
sparseness value will be used. If sparseness is manually set
this flag decides if the manual sparseness will be checked
with cross validations.}

\item{tstamp}{timestamp format used in LESYMAP}

\item{pThreshold}{(default=0.05) If cross validated
correlations show significance below this value
the results are considered null and an empty
map is returned.}

\item{mycoption}{(default=1) SCCAN parameter, see \code{\link[ANTsR]{sparseDecom2}}}

\item{robust}{(ddefault=1) SCCAN parameter, see \code{\link[ANTsR]{sparseDecom2}}}

\item{sparseness}{(default=1) SCCAN parameter, see \code{\link[ANTsR]{sparseDecom2}}}

\item{nvecs}{(default=1) SCCAN parameter, see \code{\link[ANTsR]{sparseDecom2}}}

\item{cthresh}{(default=150) SCCAN parameter, see \code{\link[ANTsR]{sparseDecom2}}}

\item{its}{(default=20) SCCAN parameter, see \code{\link[ANTsR]{sparseDecom2}}}

\item{npermsSCCAN}{(default=0) SCCAN permutations, see \code{\link[ANTsR]{sparseDecom2}}}

\item{smooth}{(default=0.4) SCCAN parameter, see \code{\link[ANTsR]{sparseDecom2}}}

\item{...}{other arguments received from \code{\link{lesymap}}.}
}
\value{
List of objects returned:
\itemize{
\item\code{statistic} - vector of statistical values
\item\code{pvalue} - vector of pvalues
\item\code{optimalSparseness} - (optional) optimal value found for sparseness
\item\code{CVcorrelation.stat} - (optional) Correlation between
true and predicted score with k-fold validation using
the optimal sparseness value
\item\code{CVcorrelation.pval} - (optional) p-value of the above correlation
}

List with vectors of \code{statistic}, \code{pvalue}, \code{optimalSparseness},
\code{CVcorrelation.stat}, \code{CVcorrelation.pval}
}
\description{
Multivariate SCCAN adapted for lesion to symptom mapping purposes.
By default an optimization routine is used to find the best
\code{sparseness} value. If you specify sparseness manually, it
will be validated to find the cross-validated correlation that
can be obtained with that sparseness. You can skip the entire
optimization/validation by choosing \code{optimizeSparseness=FALSE}.
To understand SCCAN arguments, see \code{\link[ANTsR]{sparseDecom2}}.
}
\author{
Dorian Pustina
}
