% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_svr.R
\name{lsm_svr}
\alias{lsm_svr}
\title{Support Vector Regression for symptom mapping}
\usage{
lsm_svr(lesmat, behavior, SVR.nperm = 10000,
  SVR.type = "eps-regression", SVR.kernel = "radial", SVR.gamma = 5,
  SVR.cost = 30, SVR.epsilon = 0.1, showInfo = TRUE, ...)
}
\arguments{
\item{lesmat}{matrix of voxels (columns) and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{SVR.nperm}{(default=10,000) number of permutations to run to
estimate p-values. Note, these p-values are uncorrected for
multiple comparisons.}

\item{SVR.type}{(default='eps-regression') type of SVM to run, see
\code{\link[e1071]{svm}}.}

\item{SVR.kernel}{(default='radial') type of kernel to use, see
\code{\link[e1071]{svm}}}

\item{SVR.gamma}{(default=5) gamma value, see
\code{\link[e1071]{svm}}.}

\item{SVR.cost}{(default=30) cost value, see
\code{\link[e1071]{svm}}.}

\item{SVR.epsilon}{(default=0.1) epsilon value, see
\code{\link[e1071]{svm}}.}

\item{showInfo}{logical (default=TRUE) display messages}

\item{...}{other arguments received from \code{\link{lesymap}}.}
}
\value{
List of objects returned:
\itemize{
  \item\code{statistic} - vector of statistical values
  \item\code{pvalue} - vector of pvalues
}
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
The SVR method is used. The function relies on the
\code{\link[e1071]{svm}} function of the e1071 package. The
analysis follows a similar logic found in the SVR-LSM code published
by \href{https://www.ncbi.nlm.nih.gov/pubmed/25044213}{Zhang (2015)}.
After a first run of SVM, p-values are established with a
permutation procedures as the number  of times weights are randomly
exceeded in permutations. The returned p-values are not corrected
for multiple comparisons.
}
\author{
Daniel Wiesen, Dorian Pustina
WHAT IS THE RATIONALE FOR SCALING BY 10?
WHAT IS THE RATIONALE FOR SCALING BY 10?
}
