% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_ttest.R
\name{lsm_ttest}
\alias{lsm_ttest}
\title{lsm_ttest}
\usage{
lsm_ttest(lesmat, behavior, var.equal = T, alternative = "greater", ...)
}
\arguments{
\item{lesmat}{binary matrix (0/1) of voxels (columns)
and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{var.equal}{logical (default=TRUE) should the
variance between groups considered equal (t-test) or
unequal (Welch test).}
}
\value{
List with vectors of statistic, pvalue, and zscore.
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
T-tests are performed using each column of the
matrix to split the behavioral scores in two groups. If
var.equal=TRUE the Welch test is performed instead.
}
\author{
Dorian Pustina
}

