% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minSegDistance.R
\name{minSegDistance}
\alias{minSegDistance}
\title{Metric displacement of binary masks}
\usage{
minSegDistance(manual, predict, get = "all", binarize = F, label = 1)
}
\arguments{
\item{manual}{manual segmentation of class antsImage,
used as reference}

\item{predict}{other antsImage to compare to manual}

\item{get}{(default='all') one of 'mean', 'max', 'min', or 'all'}

\item{binarize}{logical (default=FALSE) whether to binarize
the input images}

\item{label}{(default=1) integer or vector of labels to binarize
I.e., label=c(2,4) means label 2 from manual, and 4 from predict
will be compared.}
}
\value{
Scalar (for 'mean', 'max', 'min') or list (for 'all').
Note, results are in milimeters
}
\description{
This function computes the metric displacement between two binary masks
}
\note{
max = Hausdorff distance
}
\author{
Dorian Pustina
}
