% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lesymap.R
\name{print.lesymap}
\alias{print.lesymap}
\title{Console print of lesymap object}
\usage{
\method{print}{lesymap}(x, ...)
}
\arguments{
\item{x}{the output from a lesymap() call.}

\item{...}{useless for compatibility with default print.}
}
\description{
Funciton to display some meaningful summary
when a lesymap output is called in command line.
}
\author{
Dorian Pustina
}
