% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printInfo.R
\name{printInfo}
\alias{printInfo}
\title{printInfo}
\usage{
printInfo(message, type = "full", tstamp = "\%H:\%M:\%S", ...)
}
\arguments{
\item{message}{character binary matrix (0/1) of voxels (columns)
and subjects (rows).}

\item{type}{decides how to add newlines and timestamp by choosing
between the types of messages LESYMAP needs.
Choose between "full", "head", "middle", "tail".}

\item{tstamp}{the timestamp format to display}

\item{...}{other arguments that are passed by upstream functions}
}
\description{
Displays in the R console the information sent
by LESYMAP functions.
}
\author{
Dorian Pustina
}
