% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_beta_PMU.R
\name{run_beta_PMU}
\alias{run_beta_PMU}
\title{run_beta_PMU}
\usage{
run_beta_PMU(lesmat, behavior, ori_beta_val, maskIdx, posIdx, negIdx,
  maskDim, betaScale)
}
\arguments{
\item{lesmat}{matrix of voxels (columns) and subjects (rows).}

\item{behavior}{vector of behavioral scores.}
}
\value{
matrix with a random SVR model including beta weights
}
\description{
Run Permutations for SVR-LSM
This function permutes the labels and generates a random SVR model & beta weights
for each voxel location / iteration.The number of permutation is specified via nperm
}
\author{
Daniel Wiesen; based on SVR-LSM Scripts by Zhang et al., 2014
}
