% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.lesymap.R
\name{save.lesymap}
\alias{save.lesymap}
\title{Save the output of lesymap.}
\usage{
save.lesymap(lsm, saveDir, infoFile = "Info.txt", template = NA,
  saveTemplate = F, savePatchImages = T, plot.alpha = 0.8,
  plot.axis = 3, plot.quality = 8, outputLogFile = "outputLog.txt",
  ...)
}
\arguments{
\item{lsm}{object obtained with lesymap()}

\item{saveDir}{folder to save to, will be
created if it doesn't exist.}

\item{infoFile}{(default='Info.txt') what should
be the filename of the file with information.}

\item{template}{(default=NA) an antsImage to overlay
the results to. If the template is provided,
results will be plotted and saved as image.}

\item{saveTemplate}{(default=FALSE) should the
template image also be saved? Useful when
passing the results to a colleague.}

\item{savePatchImages}{(default=TRUE) should the patch images
be saved}

\item{plot.alpha}{see plot.antsImage}

\item{plot.axis}{see plot.antsImage}

\item{plot.quality}{see plot.antsImage}

\item{...}{other arguments to use for plot().}
}
\value{
Nothing is returned.
Files saved include resulting maps and
a descriptive file with a lot of
information about the lesymap run.
}
\description{
Function to save the output of lesymap.
}
\author{
Dorian Pustina
}
