% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{BM}
\alias{BM}
\title{Slow R-based Brunner-Munzel tests}
\usage{
BM(lesmat, behavior)
}
\arguments{
\item{lesmat}{matrix of voxels}

\item{behavior}{vector of behavior}
}
\value{
Returned list with:
\itemize{
\item\code{statistic} statistical values
\item\code{dof} degrees of freedom
}
}
\description{
Takes a binary matrix of voxels and a vector of behavior
and runs Brunner-Munzel tests on each voxel.
This function is not compiled and is slow.
}
\examples{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(123)
behavior = rnorm(100)
result = BM(lesmat, behavior)

}
\author{
Dorian Pustina
}
