% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueLesionPatches.R
\name{getUniqueLesionPatches}
\alias{getUniqueLesionPatches}
\title{Unique Lesion Patches}
\usage{
getUniqueLesionPatches(lesions.list, mask = NA, returnPatchMatrix = F,
  thresholdPercent = 0.1, binaryCheck = F, showInfo = T)
}
\arguments{
\item{lesions.list}{list of antsImages (faster) or filenames (slower)}

\item{mask}{a mask image to restrict the search for patches. Will be
automatically calculated if not provided (voxels > 0 in at
least one subject)}

\item{returnPatchMatrix}{logical, should the matrix of patches
be returned}

\item{thresholdPercent}{voxels with lesions in less than X percent of
subjects will not be considered (default 10\%)}

\item{binaryCheck}{logical, if True images will be verified that are binary}

\item{showInfo}{logical indicating whether to display information (default=T)}
}
\value{
List of objects named as follows:
 \itemize{
 \item\code{patchimg} - antsImage with every voxel assigned a patch number
 \item\code{patchimg.samples} - antsImage mask of one representative voxel
 for each patch. Can be used to extract the patchmatrix.
 \item\code{patchimg.size} - antsImage with the patch size at every voxel
 \item\code{patchimg.mask} - antsImage of full mask with all voxels. Can
 be used to put back results in combination with \code{patchindx}.
 \item\code{patchindx} - vector of patch membership for each voxel. Can
 be used to put back results in an image.
 \item\code{npatches} - number of unique patches in the image
 \item\code{nvoxels} - total number of lesioned voxels in \code{patchimg.mask}
 \item\code{patchvoxels} - vector of voxel count for each patch
 \item\code{patchvolumes} - vector of volume size for each patch
 \item\code{patchmatrix}  - matrix of patches. This is used in lesymap to
 save time when running repetitive analyses.
 }
}
\description{
Compute uniqe patches of voxels with the same
pattern of lesions in all subjects. Useful to understand
the amount of unique information in a lesion dataset.
I.e., your dataset can have 1000 subjects but still have
little spatial discrimination because all lesions are
similar
}
\examples{
\dontrun{
onefile = system.file(file.path('extdata','lesions'), 'Subject_001.nii.gz', package='LESYMAP')
niftifolder = dirname(onefile)

filenames = Sys.glob( file.path(niftifolder, 'Subject*.nii.gz'))
patches = getUniqueLesionPatches(filenames) # slower

lesions = imageFileNames2ImageList(filenames)
patches = getUniqueLesionPatches(lesions) # faster
}

}
\author{
Dorian Pustina
}
