% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueLesionPatches.R
\name{getUniqueLesionPatches}
\alias{getUniqueLesionPatches}
\title{Compute unique lesion patches}
\usage{
getUniqueLesionPatches(lesions.list, mask = NA,
  returnPatchMatrix = FALSE, thresholdPercent = 0.1,
  binaryCheck = FALSE, showInfo = TRUE)
}
\arguments{
\item{lesions.list}{list of antsImages (faster) or filenames (slower)}

\item{mask}{(default=NA) a mask image to restrict the search for patches. Will be
automatically calculated if not provided. Normally the mask restricts
the search only to voxels lesioned in >10\% of subejcts. To set this proportion
use \code{thresholdPercent}.}

\item{returnPatchMatrix}{(default=FALSE) logical, should the matrix of patches
be returned. This is used in \code{\link{lesymap}} to run the analyses.}

\item{thresholdPercent}{(default=0.1) voxels with lesions in less than
this proportion of subjects will not be considered. I.e., 0.1 = 10\%.}

\item{binaryCheck}{(default=FALSE) set this to TRUE to verify that maps are binary.}

\item{showInfo}{(default=TRUE) logical indicating whether to display information.}
}
\value{
List of objects named as follows:
 \itemize{
 \item\code{patchimg} - antsImage with every voxel assigned a patch number
 \item\code{patchimg.samples} - antsImage mask of one representative voxel
 for each patch. Can be used to extract the patchmatrix.
 \item\code{patchimg.size} - antsImage with the patch size at every voxel
 \item\code{patchimg.mask} - antsImage of the mask used to extract patches. Can
 be used to put back results when combined with \code{patchindx}.
 \item\code{patchindx} - vector of patch membership for each voxel. Can
 be used to put back results in an image.
 \item\code{npatches} - number of unique patches in the image
 \item\code{nvoxels} - total number of lesioned voxels in \code{patchimg.mask}
 \item\code{patchvoxels} - vector of voxel count for each patch
 \item\code{patchvolumes} - vector of volume size for each patch
 \item\code{patchmatrix}  - matrix of patches. This is used in lesymap to
 save time when running repetitive analyses.
 }
}
\description{
Compute uniqe patches of voxels with the same
pattern of lesions in all subjects. Useful to understand
the number of patterns that will be analyzed in a lesion dataset.
A patch is a group of voxels, not necessarily close to each
other, which have the same identical lesion pattern.
}
\examples{
lesydata = file.path(find.package('LESYMAP'),'extdata')

filenames = Sys.glob(file.path(lesydata, 'lesions', '*.nii.gz'))
patchinfo = getUniqueLesionPatches(filenames[1:10]) # slower

lesions = imageFileNames2ImageList(filenames[1:10])
patchinfo = getUniqueLesionPatches(lesions) # faster

}
\author{
Dorian Pustina
}
