% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesymap.R
\name{lesymap}
\alias{lesymap}
\title{Lesion to Symptom Mapping}
\usage{
lesymap(lesions.list, behavior, mask = NA, patchinfo = NA,
  method = "sccan", correctByLesSize = "none",
  multipleComparison = "fdr", pThreshold = 0.05, flipSign = F,
  minSubjectPerVoxel = "10\%", nperm = 1000, saveDir = NA,
  binaryCheck = FALSE, noPatch = FALSE, showInfo = TRUE, ...)
}
\arguments{
\item{lesions.list}{list of antsImages, or a vector of
filenames, or a single antsImage with 4 dimensions.}

\item{behavior}{vector of behavioral scores or filename
pointing to a file with a single column of numbers.}

\item{mask}{(default=NA) binary image to select the area
where analysis will be performed. If
not provided will be computed automatically
by thresholding the average lesion map at
\code{minSubjectPerVoxel}.}

\item{patchinfo}{(default=NA) an object obtained with the
\code{getUniqueLesionPatches} function. Useful to set if your
run repetitive analysese and want toavoid the computation of
patches each time. You can also pass the patchinfo object
obtained from a previous analysis.}

\item{method}{what analysis method to use to run, one of
'BM', 'BMfast', 'ttest', 'welch', 'regres', 'regresfast',
'regresPerm', 'sccan' (default) or 'svr',.

       \code{\link[=lsm_BM]{BM}} - Brunner-Munzel non parametric test, also
         called the Generalized Wilcoxon Test. The BM test is the
         same test used in the npm/Mricron software
         (see \href{https://www.ncbi.nlm.nih.gov/pubmed/17583985}{Rorden (2007)}).
         This method is slow, use 'BMfast" for a compiled faster analysis.

       \code{\link[=lsm_BMfast]{BMfast}} - ultrafast Brunner-Munzel with compiled
         code. BMfast can be
         combined with \code{multipleComparison='FWERperm'}
         to perform permutation based thresholding in a short time.

       \code{\link[=lsm_ttest]{ttest}} - Regular single tailed t-test.
         Variances of groups are assumed to be equal at each voxel, which may not
         be true. This is the test used in the voxbo
         software. Relies on \code{t.test} function in R. It is assumed
         that 0 voxels are healthy, i.e., higher behavioral scores.
         See the \code{alternative} parameter for inverted cases.
         (see \href{https://www.ncbi.nlm.nih.gov/pubmed/12704393}{Bates (2003)}).

       \code{\link[=lsm_ttest]{welch}} - t-test that does not assume equal
         variance between groups. Relies on \code{t.test} function in R.

       \code{\link[=lsm_regres]{regres}} - linear model between voxel values
         and behavior. Uses the \code{lm} function in R. This is equivalent to a
         t-test, but is useful when voxel values are continuous. This method is
         R-based and slow, for faster analysis and to add covariates
         use the \code{"regresfast"} method compiled in LESYMAP.

       \code{\link[=lsm_regresfast]{regresfast}} - fast linear regressions with
         compiled code. This method allows setting covariates. If covariates are
         specified the effect of each voxel will be estimated with the formula:\cr
         \code{behavior ~ voxel + covar1 + covar2 + ...}\cr
         The effect of covariates at each voxel is established with
         the Freedman-Lane method
         (see \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4010955/}{Winkler (2014)}).
         This LESYMAP method allows multiple comparison correction with permutation
         based methods \code{"FWERperm"} and \code{"clusterPerm"}.

       \code{\link[=lsm_regresPerm]{regresPerm}} - linear model between voxel values
         and behavior. The p-value of each individual voxel is established by
         permuting voxel values. The lmPerm package is used for this purpose.
         Note, these permutations do not correct for
         multiple comparisons, they only establish voxel-wise p-values, a correction
         for multiple comparisons is still required.

       \code{\link[=lsm_chisq]{chisq}} - chi-square test between voxel values and behavior.
         The method is used when behavioral scores are binary (i.e. presence of absence
         of deficit). Relies on the \code{\link[stats]{chisq.test}}
         R function. By default this method corrects individual voxel p-values with the
         Yates method (the same approach offered in the Voxbo software).

       \code{\link[=lsm_chisq]{chisqPerm}} - chi-square tests. P-values are established
         through permutation tests instead of regular statistics.
         Relies on the \code{\link[stats]{chisq.test}} R function.

       \code{\link[=lsm_sccan]{sccan}} - sparse canonical correlations. Multivariate
         method that considers all voxels at once by searching for voxel weights
         that collectively explain behavioral variance. For our purposes, this
         method can be considered a sparse regression technice. By default,
         lesymap will run a lengthy procedure to determine the optimal
         sparseness value (how extensive the results should be). You
         can set \code{optimizeSparseness=FALSE} if you want to skip this
         optimization. The search for optimal sparsness provides
         a cross-validated correlation measure that shows how well the
         sparseness value can predict new patients. If this predictive correlation
         is below significance (i.e., below \code{pThreshold}), the entire solution
         will be dropped and LESYMAP will return an empty statistical map.
         If the predictive correlation is significant, LESYMAP will return
         a statistical image with normalized weights between -1 and 1. The raw SCCAN
         weights image is returned in \code{rawWeights.img} as well. LESYMAP scales
         and centers both lesion and behavior data before running SCCAN
         (hardcoded in \link{lsm_sccan}). See more details in
         \href{https://www.ncbi.nlm.nih.gov/pubmed/28882479}{Pustina (2018)}

       \code{\link[=lsm_svr]{svr}} - support vector regression. Multivariate
         method that considers all voxels at once by searching for voxel weights.
         To establish p-values, a number of permutations are needed as set with
         \code{SVR.nperm}. The current implementation of SVR in LESYMAP is not
         parallelized and takes many hours to finish all permutations. The SVR
         method is initially described in
         \href{https://www.ncbi.nlm.nih.gov/pubmed/25044213}{Zhang (2014)},
         LESYMAP uses a contribution by the
         \href{https://www.ncbi.nlm.nih.gov/pubmed/30549154}{Tuebingen group}.}

\item{correctByLesSize}{whether to correct for lesion size in the analysis.
Options are "none", "voxel", "behavior", "both":
\itemize{
\item\code{"none"}: (default) no correction
\item\code{"voxel"}: divide voxel values by 1/sqrt(lesionsize).
   This is the method used in
   \href{https://www.ncbi.nlm.nih.gov/pubmed/25879574}{Mirman (2015)} and
   \href{https://www.ncbi.nlm.nih.gov/pubmed/25044213}{Zhang (2014)}.
   This correction works only with
   'regres' methods. Two sample comparisons
   (t-tests and Brunner-Munzel) use binary voxels
   and will ignore this correction.
\item\code{"behavior"}: residualize behavioral scores by removing
   the effect of lesion size. This works on all methods,
   but is more agressive on results.
\item\code{"both"}: both voxel and behavior residualized.
   }}

\item{multipleComparison}{(default='fdr') method to adjust p-values.
Standard methods include \code{"holm"}, \code{"hochberg"},
\code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}.
(see \code{\link[stats]{p.adjust}} ) \cr
Permutation methods include: \cr
\code{"FWERperm"} (permutation based family-wise threshold) is enabled
  with methods 'BMfast' and 'regresfast'. In this
  case, many analysis are run with permuted behavioral
  scores, and the peak score is recorded each time (see
  Winkler 2014). The optimal threshold is established at
  95th percentile of this distribution (or whatever pThreshold
  you choose). You can choose to use as reference another
  voxel lower in the ranks by specifying another `v` value
  (i.e., lesymap(..., v=10) will record the 10th highest voxel). \cr
\code{"clusterPerm"} (permutation based cluster correction) is enabled
  for 'regresfast'. It records the maximal
  cluster size from many random permutations of the behavior
  score and sets a cluster threshold based on that distribution.
  You must select pThreshold (voxel-wise, default=0.05) and
  clusterPermThreshold (cluster-wise, default 0.05) to achieve
  optimal thresholding with this approach.}

\item{pThreshold}{(default=0.05) threshold statistics at this p-value
(after corrections or permutations)}

\item{flipSign}{logical (default=FALSE), invert the sign in the statistics image.}

\item{minSubjectPerVoxel}{(default='10\%') remove voxels/patches with lesions
in less than X subjects. Value can
be speficifed as percentage ('10\%')
or exact number of subjects (10).}

\item{nperm}{(default=1000) number of permutations to run when
necessary. This is used mostly for univariate analyses, while
multivariate methods have their own permutation arguments.
Check the documentation of each method to know more.}

\item{saveDir}{(default=NA) save results in the specified folder.}

\item{binaryCheck}{logical (default=FALSE), make sure the lesion matrix is 0/1.
This will help if lesion maps are drawn in MRIcron or other software which
label lesioned voxel with value 255.}

\item{noPatch}{logical (default=FALSE), if True avoids using patch information
and will analyze all voxels. It will take longer and results will be
worse due to more multiple comparison corrections. This argument
is ignored when performing multivariate analyses, SCCAN or SVR, for
which all voxels are always used.}

\item{showInfo}{logical (default=TRUE), display time-stamped info messages}

\item{...}{arguments that will be passed down to other functions
(i.e., sparsness=0.045)}
}
\value{
The following objects are typically found in the returned list:
\itemize{
\item\code{stat.img}   - statistical map
\item\code{rawWeights.img}   - (optional) raw SCCAN weights
\item\code{pval.img}   - (optional) p-values map
\item\code{zmap.img}   - (optional) zscore map
\item\code{mask.img}    - mask used for the analyses
\item\code{average.img} - map of all lesions averaged,
           produced only if no mask is defined.
\item\code{callinfo} - list of details of how you called lesymap
\item\code{outputLog} - terminal output in a character variable
\item\code{perm.vector} - (optional) the values obtained from each permutation
\item\code{perm.clusterThreshold} - (optional) threshold computed
           for cluster thresholding
\item\code{perm.FWERthresh} - (optional) threshold computed for FWERperm
           thresholding
\item\code{patchinfo}   - list of variables describing patch information:
    \itemize{
    \item\code{patchimg} - antsImage with the patch number each voxels belongs to
    \item\code{patchimg.samples} - antsImage mask with a single voxel per patch
    \item\code{patchimg.size} - antsImage with the patch size at each voxel
    \item\code{patchimg.mask} -  the mask within which the function will look for patches
    \item\code{npatches} - number of unique patches in the image
    \item\code{nvoxels} - total number of lesioned voxels in mask
    \item\code{patchvoxels} - vector of voxel count for each patch
    \item\code{patchvolumes} - vector of volume size for each patch
    \item\code{patchmatrix} - the lesional matrix, ready for use in analyses.
           Matrix has size NxP (N=number of subjects, P=number of
           patches)
           }
}
}
\description{
Lesymap uses univariate and multivariate methods to map
functional regions of the brain that, when lesioned,
cause specific cognitive deficits. It requires is
a set of binary lesion maps (nifti files) in template
space and the vector of corresponding behavioral scores.
Note, lesions must be already registered in template space,
use the built-in function
\code{\link[=registerLesionToTemplate]{registerLesionToTemplate}}
or other ANTs tools to register lesions. Lesymap will check that
lesion maps are in the same space before running. For traditional
mass-univariate analyses (i.e., BMfast, ttest, etc.),
voxels with identical lesion patterns are grouped together in
unique patches. Patch-based analysis decreases the number of
multiple comparisons and speeds up the analyses. Multivariate
analysis are performed using an optimized version of sparse
canonical correlations (SCCAN, \code{method='sccan'})or support
vector regression (\code{method='svr'}).
}
\details{
Several other parameters can be specified to lesymap()
which will be passed to other called fuctions. Here are
some examples:

\code{permuteNthreshold}  - (default=9) for Brunner-Munzel tests
   in \code{method='BMfast'} or \code{method='BM'}.
   Voxels lesioned in less than this number
   of subjects will undergo permutation-based
   p-value estimation. Useful because the BM test
   is not valid when comparing groups with N < 9.
   Issue described in:
   \href{https://www.ncbi.nlm.nih.gov/pubmed/19766664}{Medina (2010)}

\code{clusterPermThreshold} - threshold used to find the optimal cluster size when
   using \code{multipleComparison='clusterPerm'}.

\code{alternative} - (default='greater') for two sample tests (ttests and BM).
   By default LESYMAP computes single tailed p-values assuming
   that non-lesioned 0 voxels have higher behavioral scores.
   You can specify the opposite relationship with alternative='less'
   or compute two tailed p-values with alternative='two.sided'.

\code{covariates} - (default=NA) enabled for method = 'regresfast'.
   This will allow to model the effect of each voxel
   in the context of other covariates, i.e., formula
   "behavior ~ voxel + covar1 + covar2 + ...". \cr
   I,.e., lesymap(lesions,behavior, method='regresfast',
           covariates=cbind(lesionsize, age)).\cr
   If you choose permutation based thresholding with covariates, lesymap
   will use the Freedman-Lane method for extracting the unique effect of
   each voxel (see Winkler 2014, Freedman 1983)

\code{template} - antsImage or filename used for plotting the results if a saving
   directory is specified (see \code{saveDir})

\code{v} - (default=1) which voxel to record for permutation based thresholding.
   Normally the peak voxel is used (1), but other voxels can be recorded.
   See \href{https://www.ncbi.nlm.nih.gov/pubmed/28847712}{Mirman (2017)}
   for this approach.
}
\examples{
lesydata = file.path(find.package('LESYMAP'),'extdata')
filenames = Sys.glob(file.path(lesydata, 'lesions', 'Subject*.nii.gz'))
behavior = Sys.glob(file.path(lesydata, 'behavior', 'behavior.txt'))
template = antsImageRead(
 Sys.glob(file.path(lesydata, 'template', 'ch2.nii.gz')))
lsm = lesymap(filenames, behavior, method = 'BMfast')
plot(template, lsm$stat.img, window.overlay = range(lsm$stat.img))

\dontrun{
# Same analysis with SCCAN
lsm = lesymap(filenames, behavior, method = 'sccan',
sparseness=0.045, validateSparseness=FALSE)
plot(template, lsm$stat.img, window.overlay = range(lsm$stat.img))
save.lesymap(lsm, saveDir='/home/dp/Desktop/SCCANresults')
}

}
\author{
Dorian Pustina
}
