% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_BM.R
\name{lsm_BM}
\alias{lsm_BM}
\title{Brunner-Munzel tests for symptom mapping (slow)}
\usage{
lsm_BM(lesmat, behavior, permuteNthreshold = 9, nperm = 10000,
  alternative = "greater", showInfo = TRUE, ...)
}
\arguments{
\item{lesmat}{binary matrix (0/1) of voxels (columns)
and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{permuteNthreshold}{(default=9) Voxels lesioned in less than
this number will undergo permutation based thresholding.
See Medina et al 2010.}

\item{nperm}{Number of permutations to perform when needed.}

\item{alternative}{(default="greater") It is assumed that
healthy voxels (0) have greater behavioral scores. If your
data follow an inverted relationship choose "less" or
"two.sided".}

\item{showInfo}{display info messagges when running the function.}

\item{...}{other arguments received from \code{\link{lesymap}}.}
}
\value{
List of objects returned:
\itemize{
 \item\code{statistic} - vector of statistical values
 \item\code{pvalue} - vector of pvalues
 \item\code{zscore} - vector of zscores
}
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
Brunner-Munzel tests are performed using each column
of the matrix to split the behavioral scores in two
groups.
}
\examples{
{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(123)
behavior = rnorm(100)
result = lsm_BM(lesmat, behavior)
}

}
\author{
Dorian Pustina
}
