% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_regres.R
\name{lsm_regres}
\alias{lsm_regres}
\title{Regression tests for symptom mapping (slow)}
\usage{
lsm_regres(lesmat, behavior)
}
\arguments{
\item{lesmat}{matrix of voxels (columns) and subjects (raws).}

\item{behavior}{vector of behavioral scores.}
}
\value{
List of objects returned:
\itemize{
\item\code{statistic} - vector of statistical values
\item\code{pvalue} - vector of pvalues
\item\code{zscore} - vector of zscores
}
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
Regressions are performed between behavior and each column in the
lesmat matrix.
}
\examples{
{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(123)
behavior = rnorm(100)
result = lsm_regres(lesmat, behavior)
}

}
\author{
Dorian Pustina
}
