% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm_ttest.R
\name{lsm_ttest}
\alias{lsm_ttest}
\title{lsm_ttest}
\usage{
lsm_ttest(lesmat, behavior, var.equal = T, alternative = "greater",
  checkAssumptions = TRUE, showInfo = TRUE, ...)
}
\arguments{
\item{lesmat}{binary matrix (0/1) of voxels (columns)
and subjects (rows).}

\item{behavior}{vector of behavioral scores.}

\item{var.equal}{logical (default=TRUE) should the
variance between groups considered equal (t-test) or
unequal (Welch test).}

\item{alternative}{(default='greater') Sets the expected
relationship between voxel value and behavior. By default
voxels with zero are not lesioned, and behavior is expected to
be higher, thus \code{alternative='greater'}. If the relationship in your
data is inverted, use \code{alternative='less'}, and if
you don't have a relationship hypothesis data,
use \code{alternative='two.sided'}.}

\item{checkAssumptions}{Check whether t-test assumptions are met
for every voxel}

\item{showInfo}{logical (default=TRUE), display time-stamped info messages}

\item{...}{other arguments received from \code{\link{lesymap}}.}
}
\value{
List of objects returned:
\itemize{
\item\code{statistic} - vector of statistical values
\item\code{pvalue} - vector of pvalues
\item\code{zscore} - vector of zscores
}
}
\description{
Lesion to symptom mapping performed on a prepared matrix.
T-tests are performed using each column of the
matrix to split the behavioral scores in two groups. If
var.equal=TRUE the Welch test is performed instead.
}
\examples{
{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(123)
behavior = rnorm(100)
result = lsm_ttest(lesmat, behavior)
}

}
\author{
Dorian Pustina
}
