% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_SCCANsparseness.R
\name{optimize_SCCANsparseness}
\alias{optimize_SCCANsparseness}
\title{Optimization of SCCAN sparseness}
\usage{
optimize_SCCANsparseness(lesmat, behavior, mask, nfolds = 4,
  sparsenessPenalty = 0.03, lower = 0.005, upper = 0.9, tol = 0.03,
  justValidate = F, cvRepetitions = 3, mycoption = 1, robust = 1,
  sparseness = NA, nvecs = 1, cthresh = 150, its = 30,
  npermsSCCAN = 0, smooth = 0.4, sparseness.behav = -0.99, ...)
}
\arguments{
\item{lesmat}{lesion matrix}

\item{behavior}{behavior vector}

\item{mask}{antsImage mask}

\item{nfolds}{how many folds to use}

\item{sparsenessPenalty}{penalty term}

\item{lower}{minimum searched sparseness}

\item{upper}{maximum searched sparseness}

\item{tol}{tolerance value, see optimize() in R}

\item{justValidate}{just check the CV of provided sparseness}

\item{cvRepetitions}{number of cross-validations}

\item{mycoption}{standard SCCAN parameter}

\item{sparseness}{standard SCCAN parameter}

\item{nvecs}{standard SCCAN parameter}

\item{cthresh}{standard SCCAN parameter}

\item{its}{standard SCCAN parameter}

\item{npermsSCCAN}{SCCAN permutations}

\item{smooth}{standard SCCAN parameter}

\item{sparseness.behav}{what sparsness to use for behavior}

\item{robbust}{standard SCCAN parameter}
}
\value{
List with: \cr
\code{minimum} - best sparseness value \cr
\code{objective} - minimum value of objective function \cr
\code{CVcorrelation} - cross-validated correlation of optimal sparness \cr
}
\description{
Function used to optimize SCCAN sparseness for
lesion to symptom mapping.
}
\author{
Dorian Pustina


the optimization function
Will run SCCAN on each training fold, compute
behavior prediction on the test fold, and finally
return a cross validated correlation from entire sample

end of optimfun
}

