% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_SCCANsparseness.R
\name{optimize_SCCANsparseness}
\alias{optimize_SCCANsparseness}
\title{Optimization of SCCAN sparseness}
\usage{
optimize_SCCANsparseness(lesmat, behavior, mask, nfolds = 4,
  sparsenessPenalty = 0.03, lower = 0.005, upper = 0.9, tol = 0.03,
  justValidate = FALSE, cvRepetitions = ifelse(length(behavior) <= 30,
  6, ifelse(length(behavior) <= 40, 5, ifelse(length(behavior) <= 50, 4,
  3))), showInfo = TRUE, mycoption = 1, robust = 1,
  sparseness = NA, nvecs = 1, cthresh = 150, its = 30,
  npermsSCCAN = 0, smooth = 0.4, sparseness.behav = -0.99,
  maxBased = FALSE, directionalSCCAN = FALSE, ...)
}
\arguments{
\item{lesmat}{lesion matrix}

\item{behavior}{behavior vector}

\item{mask}{antsImage mask}

\item{nfolds}{how many folds to use}

\item{sparsenessPenalty}{penalty term}

\item{lower}{minimum searched sparseness}

\item{upper}{maximum searched sparseness}

\item{tol}{tolerance value, see optimize() in R}

\item{justValidate}{just check the CV of provided sparseness}

\item{cvRepetitions}{number of cross-validations at each sparseness
value. Dynamically set depending on sample size: <=30 to 6 reps,
<=40 to 5 reps, <=50 to 4 reps, > 50 to 3 reps.}

\item{showInfo}{logical (default=TRUE) display messages}

\item{mycoption}{standard SCCAN parameter}

\item{robust}{standard SCCAN parameter}

\item{sparseness}{standard SCCAN parameter}

\item{nvecs}{standard SCCAN parameter}

\item{cthresh}{standard SCCAN parameter}

\item{its}{standard SCCAN parameter}

\item{npermsSCCAN}{SCCAN permutations}

\item{smooth}{standard SCCAN parameter}

\item{sparseness.behav}{what sparsness to use for behavior}

\item{maxBased}{standard SCCAN parameter}

\item{directionalSCCAN}{(default=FALSE) If TRUE, the upper and lower
bounds of sparseness search will be negative. A negative sparseness
permits positive and negative voxel weights, thus finding the
direction of the relationship with behavior.}

\item{...}{other arguments received from \code{\link{lesymap}}
or \code{\link{lsm_sccan}}.}
}
\value{
List with: \cr
\code{minimum} - best sparseness value \cr
\code{objective} - minimum value of objective function \cr
\code{CVcorrelation} - cross-validated correlation of optimal sparness \cr
}
\description{
Function used to optimize SCCAN sparseness for
lesion to symptom mapping.
}
\author{
Dorian Pustina


the optimization function
Will run SCCAN on each training fold, compute
behavior prediction on the test fold, and finally
return a cross validated correlation from entire sample

end of optimfun
}
