% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{regresfast}
\alias{regresfast}
\title{Fast linear regressions}
\usage{
regresfast(X, y, covariates, hascovar = FALSE)
}
\arguments{
\item{X}{matrix of voxlels (columns) for all
subjects (rows).}

\item{y}{vector of behavioral scores.}

\item{covariates}{matrix with one or more columns.
Must be of same length as behavior. This variable
should always be set, and the next argument can tell
if covariates should be used or not.}

\item{hascovar}{logical to tell whether covariates
should be used.}
}
\value{
List with (1) statistic, (2) n = number of
subjects, (3) kxfm = degrees of freedom.
}
\description{
Takes a matrix of voxels and a vector of behavior
and runs fast regressions for each voxel. Covariates
can be defined (i.e. age) to find the effect of each
voxel on behavior within the context of other predictive
factors.
}
\author{
Dorian Pustina
}

